chore(ci): Automatically upload the database at the end of the deploy phase
The database is located on data/export-data-firebase.json. All deploy process will be followed by a full upload of the database.
A check operation of the database is executed during the script phase to validate the json with the took JQ. If the json is invalid, this will stop the build phase.
This step of validation will allow to check this file for PR process.

Closes #75